LayoutTests/imported/w3c/web-platform-tests/css/css-color/parsing/color-valid-hwb-expected.txt


PASS e.style['color'] = "hwb(120 30% 50%)" should set the property value
PASS e.style['color'] = "hwb(120 30% 50% / 0.5)" should set the property value
PASS e.style['color'] = "hwb(none none none)" should set the property value
PASS e.style['color'] = "hwb(0 0% 0%)" should set the property value
PASS e.style['color'] = "hwb(none none none / none)" should set the property value
PASS e.style['color'] = "hwb(0 0% 0% / 0)" should set the property value
PASS e.style['color'] = "hwb(120 none none)" should set the property value
PASS e.style['color'] = "hwb(120 0% 0%)" should set the property value
PASS e.style['color'] = "hwb(120 80% none)" should set the property value
PASS e.style['color'] = "hwb(120 80% 0%)" should set the property value
PASS e.style['color'] = "hwb(120 none 50%)" should set the property value
PASS e.style['color'] = "hwb(120 0% 50%)" should set the property value
PASS e.style['color'] = "hwb(120 30% 50% / none)" should set the property value
PASS e.style['color'] = "hwb(120 30% 50% / 0)" should set the property value
PASS e.style['color'] = "hwb(none 100% 50% / none)" should set the property value
PASS e.style['color'] = "hwb(0 100% 50% / 0)" should set the property value
FAIL e.style['color'] = "hwb(120 30 50)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30 50 / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30% 50)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30% 50 / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30 50%)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30 50% / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 none 50)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 none 50 / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30 none)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30 none / 0.5)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(120 30 50 / none)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['color'] = "hwb(320deg 30% 40%)" should set the property value assert_equals: serialization should be canonical expected "rgb(153, 77, 128)" but got "rgb(153, 77, 127)"
FAIL e.style['color'] = "hwb(calc(infinity) 20% 10%)" should set the property value assert_equals: serialization should be canonical expected "rgb(230, 51, 51)" but got "rgb(51, 51, 51)"
FAIL e.style['color'] = "hwb(calc(-infinity) 20% 10%)" should set the property value assert_equals: serialization should be canonical expected "rgb(230, 51, 51)" but got "rgb(51, 51, 51)"
PASS e.style['color'] = "hwb(calc(0 / 0) 20% 10%)" should set the property value
PASS e.style['color'] = "hwb(90 20% 10% / calc(infinity))" should set the property value
PASS e.style['color'] = "hwb(90 20% 10% / calc(-infinity))" should set the property value
PASS e.style['color'] = "hwb(90 20% 10% / calc(0 / 0))" should set the property value
